requirements.txt
dash_core_components 
dash_html_components 
dash_bootstrap_components 
dash
dash_table
plotly
flask
pandas 
numpy 
feedparser
requests
folium
newsapi-python
numpy
dash_leaflet